+++ src/components/ItaliaTheme/Blocks/ArgumentsInEvidence/Edit.jsx
@@ -2,19 +2,58 @@
 import React from 'react'
 import BodyWrapper from './BodyWrapper'
 import PropTypes from 'prop-types';
+import BottomBody from './BottomBody'
+import Body from './Body'
+import { withDNDContext, SubblocksEdit, SubblocksWrapper } from 'volto-subblocks'
+import { SidebarPortal } from '@plone/volto/components'
+import Sidebar from './Sidebar.jsx'
 
-const Edit = ({ data, id, block, selected, intl, onChangeBlock }) => {
-  return (
-    <BodyWrapper 
-      data={data}
-      inEditMode={true}
-      draggable={true}
-      block={block}
-      onChangeBlock={onChangeBlock}
-      selected={selected}
-      intl={intl}
-    />
-  )
+class Edit extends SubblocksEdit {
+  render() {
+    return (
+      <div className='argumentInEvidence public-ui'>
+        <SubblocksWrapper node={this.node}>
+          <BodyWrapper
+            data={this.props.data}
+            inEditMode={false}
+          >
+            {this.state.subblocks.map((subblock, subindex) => (
+              <Body
+                data={subblock}
+                index={subindex}
+                key={subblock.id}
+                inEditMode={true}
+                selected={this.isSubblockSelected(subindex)}
+                {...this.subblockProps}
+                intl={this.props.intl}
+                onChangeBlock={this.props.onChangeBlock}
+                openObjectBrowser={this.props.openObjectBrowser}
+                onChangeFocus={this.onSubblockChangeFocus}
+              />
+            ))}
+            {this.props.selected && this.renderAddBlockButton()}
+          </BodyWrapper>
+        </SubblocksWrapper>
+
+        <SidebarPortal selected={this.props.selected || false}>
+          <Sidebar
+            {...this.props}
+            data={this.props.data}
+            block={this.props.block}
+            onChangeBlock={this.props.onChangeBlock}
+            onChangeSubBlock={this.onChangeSubblocks}
+            selected={this.state.subIndexSelected}
+            setSelected={this.onSubblockChangeFocus}
+          />
+        </SidebarPortal>
+        
+        <BottomBody 
+          data={this.props.data}
+          intl={this.props.intl}
+        />
+      </div>
+    )
+  }
 }
 /**
  * Property types.
@@ -30,5 +69,5 @@ Edit.propTypes = {
   onChangeBlock: PropTypes.func.isRequired,
 };
 
-export default Edit;
+export default React.memo(withDNDContext(Edit));;
